rename manybody_jastrow to j3
